+++ lib/gg_publish.dart
@@ -8,5 +8,6 @@ library;
 
 export 'src/gg_publish.dart';
 export 'src/commands/is_published.dart';
+export 'src/commands/is_latest_state_published.dart';
 export 'src/commands/is_upgraded.dart';
 export 'src/commands/publish.dart';