+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Rework changelog
 - 'Github Actions Pipeline'
 - 'Github Actions Pipeline: Add SDK file containing flutter into .github/workflows to make github installing flutter and not dart SDK'
+- BREAKING CHANGE:  is renamed into . Added a new  which returns true when the package is published at all
 
 ## 2.0.3 - 2024-01-01